fix(save): ensure module's save method functions
Added a #save method to baseResourceConveter, which invokes
fileExporter.save. 😒

GH-16
