+++ .npmignore
@@ -1,0 +1,35 @@
+.bowerrc
+.cssTmp/
+.editorconfig
+.gitattributes
+.gitignore
+.gulp-scss-cache/
+.idea/
+.jshintrc
+.jsTmp/
+.nyc_output
+.sass-cache/
+.scannerwork
+.travis.yml
+.vscode
+*.csv
+appveyor.yml
+bower.json
+bower/
+coverage/
+docs/
+example/
+gulp/
+gulpfile.js
+karma-coverage.conf.js
+karma.conf.js
+reports
+src/
+test/
+tsconfig.json
+tsd.json
+tslint.json
+typings/
+lib/__fixtures__/
+lib/__mocks__/
+lib/__tests__/
