+++ .npmrc
@@ -1,0 +1,2 @@
+package-lock=false
+save-exact=true
