+++ lib/__tests__/base-resource-converter.test.js
@@ -1,10 +1,12 @@
+const {noop} = require('lodash')
 const allIssuesPrs = require('../__fixtures__/all-issues-prs')
 const baseResourceConverter = require('../base-resource-converter')
+const defaultOptions = require('../default-options')
+const fileExporter = require('../file-exporter')
 const githubResourceLogger = require('../github-resource-logger')
 const issuesGetForRepoResponse = require('../__fixtures__/issues/get-for-repo-response')
 const Octokit = require('@octokit/rest')
 const prsGetAllResponse = require('../__fixtures__/pull-requests/get-all-response')
-const defaultOptions = require('../default-options')
 const restClientOptions = require('../rest-client-options')
 
 describe('baseResourceConverter', () => {
@@ -152,6 +154,14 @@ describe('baseResourceConverter', () => {
       const noopMock = jest.fn()
       baseResourceConverter.paginate(noopMock)
       expect(noopMock).toHaveBeenCalledWith(defaultOptions.api)
+    })
+  })
+
+  describe('#save', () => {
+    it('writes data to the file system', () => {
+      spyOn(fileExporter, 'save').and.callFake(noop).and.callThrough()
+      baseResourceConverter.save()
+      expect(fileExporter.save).toHaveBeenCalled()
     })
   })
 })
