+++ lib/__tests__/index.test.js
@@ -47,6 +47,7 @@ describe('index', () => {
 
   describe('#save', () => {
     it('references baseResourceConverter#save', () => {
+      expect(index.save).toBeDefined()
       expect(index.save).toBe(baseResourceConverter.save)
     })
   })
