+++ lib/base-resource-converter.js
@@ -2,7 +2,7 @@ const dotenv = require('dotenv')
 const githubResourceLogger = require('./github-resource-logger')
 const Octokit = require('@octokit/rest')
 const defaultOptions = require('./default-options')
-
+const fileExporter = require('./file-exporter')
 const envConfig = dotenv.config()
 const octokit = new Octokit(defaultOptions.api)
 
@@ -98,7 +98,11 @@ const baseResourceConverter = {
 
   getPullRequestsForRepo,
 
-  paginate
+  paginate,
+
+  save (data, dest) {
+    fileExporter.save(dest, data)
+  }
 }
 
 module.exports = baseResourceConverter
