+++ lib/default-options.js
@@ -1,3 +1,4 @@
+const enUs = require('./templates/help/en-us')
 const meow = require('meow')
 const restClientOptions = require('./rest-client-options')
 
@@ -55,7 +56,7 @@ const cliOptions = {
   }
 }
 
-const meta = meow(cliOptions)
+const meta = meow(enUs.help, cliOptions)
 
 /**
  * Input options, especillay for the CLI.
