+++ lib/file-exporter.js
@@ -55,9 +55,8 @@ const decorateFileName = (options = defaultOptions.meta.flags) => {
 
 const writeToFile = async (dest, data) => {
   try {
-    console.log('Exporting...')
     await fsExtra.outputFile(dest, data)
-    console.log('Saved "%s".', dest)
+    githubResourceLogger.info(`Saved "${dest}".`)
   } catch (err) {
     githubResourceLogger.error(err)
   }
