+++ lib/github-resource-logger.js
@@ -11,7 +11,7 @@ const outputModeEnum = ((enumerator) => {
 })({
 })
 
-const outputFormat = bunyanFormat({
+const formatOutput = bunyanFormat({
   'color': true,
   'outputMode': outputModeEnum.LONG
 })
@@ -29,7 +29,7 @@ const outputFormat = bunyanFormat({
 const githubResourceLogger = bunyan.createLogger({
   'level': bunyan.INFO,
   'name': 'github-resource-converter',
-  'stream': outputFormat
+  'stream': formatOutput
 })
 
 /**
