+++ lib/index.js
@@ -3,7 +3,7 @@ const csvMediator = require('./csv-mediator')
 const defaultOptions = require('./default-options')
 const githubResourceLogger = require('./github-resource-logger')
 
-module.exports = {
+const githubResourceConverter = {
   /**
    * Synchronously establish your account or organizational identity with GitHub.
    *
@@ -33,7 +33,9 @@ module.exports = {
     'getForRepo': baseResourceConverter.getPullRequestsForRepo
   },
 
-  'save': baseResourceConverter.save,
+  'save': csvMediator.save,
 
   'toCsv': csvMediator.toCsv
 }
+
+module.exports = githubResourceConverter
